PROG_GameMode_Fix_Replicated Variables And Functions
MainFlag and MiniFlag blueprints set to replicate
HUD Widget gets all of its data from the GameState now, timer, rounds and team points are properly showing up
GameState TMap of teams changed to TArray of teams because Maps were not compatible with Unreal Engin networking
Since listOfTeams is now an array access the team you want with static_cast<int32>(ETeamIdentifier::TeamName)
GameState, all variables set to replicate
GameState
GameMode AddPoints function delegate set per player in the PostLogin function
GameMode SpawnPlayer functions created (still need to be implemented)
GameMode updates timer every second instead of OnTick
GameMode Mercy rule check fixed, forgot to dereference the pointers
Team replicates points variable
TeamIdentifier order of values changed (Human, Alien, None instead of None, Human, Alien)
PlayerState all variables replicated
Removed BP_Character_PEN

+++ Source/GAME259A/BaseCharacter.cpp
@@ -1,6 +1,7 @@
 #include "BaseCharacter.h"
 #include "Kismet/KismetMathLibrary.h"
 #include "GAME259A/GameMode/CTFGameState.h"
+#include "GameMode/Team.h"
 #include "CTFPlayerState.h"
 
 // Sets default values
@@ -158,7 +159,7 @@ void ABaseCharacter::Respawn()
 		{
 			if(gameState->listOfTeams.Num() != 0)
 			{
-				gameState->listOfTeams[ctfPlayerState->teamID]->SpawnPlayer(this);
+				gameState->listOfTeams[static_cast<int32>(ctfPlayerState->teamID)]->SpawnPlayer(this);
 			} else
 			{
 				UE_LOG(LogTemp, Warning, TEXT("No teams found in current level so respawning will not be handled, Try testing in the GameModeTestMap"));
